WIP Sort out applet repainting
50fps, double-buffered, to off-screen Graphics; frames skipped if no change.
Blitted _asynchronously_ to screen (regardless of update rate).

Tweaks to CDasherInterfaceBase repaint logic. Overriding Redraw() to schedule
 a call to NewFrame (as only source of calls to NewFrame) now a valid technique.

Also change CDasherScreen::DrawString() to take size as int not long (!)

+++ src/dasher/CDasherScreen.java
@@ -71,7 +71,7 @@ public interface CDasherScreen {
 	 * @param y1 Top-left hand corner y co-ordinate
 	 * @param Size Font size in points
 	 */
-	public abstract void DrawString(String string, int x1, int y1, long Size);
+	public abstract void DrawString(String string, int x1, int y1, int Size);
 	
   /** Draw a filled rectangle
 	*
